Refactor QuestionService for Brevity
